Fix no_offset_view for Slice (#376)
Currently, the implementation of `no_offset_view` for a `Slice` doesn't
actually ensure that there is no offset. This is because
`Base.Slice(UnitRange(a))` has `UnitRange(a)` as its axes, which are
usually offset. This PR changes the behavior to return a `UnitRange(a)`
instead.

Fixes #375 

The flip side of the current implementation is that the `Slice`
information is lost, so `IndexStyle` of 2D views might become
`IndexCartesian` once they are routed through `no_offset_view`.

```julia
julia> arr = OffsetArray(reshape(1:15, 3, 5), 2, 3)
3×5 OffsetArray(reshape(::UnitRange{Int64}, 3, 5), 3:5, 4:8) with eltype Int64 with indices 3:5×4:8:
 1  4  7  10  13
 2  5  8  11  14
 3  6  9  12  15

julia> V = view(arr, :, :)
3×5 view(OffsetArray(reshape(::UnitRange{Int64}, 3, 5), 3:5, 4:8), :, :) with eltype Int64 with indices 3:5×4:8:
 1  4  7  10  13
 2  5  8  11  14
 3  6  9  12  15

julia> IndexStyle(V)
IndexLinear()

julia> V_nooff = OffsetArrays.no_offset_view(V)
3×5 view(OffsetArray(reshape(::UnitRange{Int64}, 3, 5), 3:5, 4:8), 3:5, 4:8) with eltype Int64:
 1  4  7  10  13
 2  5  8  11  14
 3  6  9  12  15

julia> IndexStyle(V_nooff)
IndexCartesian()
```

An alternate approach might be to shift the parent to one-based indices,
and use `Base.Slice{Base.OneTo{Int}}` axes, which preserve this
information. This, however, departs from the current approach of
preserving the `parentindices` of the `SubArray`.

Edit: the alternate approach is implemented now.